removed old api functions
