+++ apis/candidateApis.js
@@ -20,7 +20,8 @@ const { sanitize,
 const candidateApis = {
     POST_updateAllOnboarding,
     POST_candidate,
-    POST_endOnboarding
+    POST_endOnboarding,
+    POST_sendVerificationEmail
 }
 
 
@@ -307,6 +308,50 @@ function POST_endOnboarding(req, res) {
         } else {
             res.status(500).send("Error ending onboarding.");
         }
+    });
+}
+
+
+function POST_sendVerificationEmail(req, res) {
+    let email = sanitize(req.body.email);
+    let query = {email: email};
+
+    let moonshotUrl = 'https://www.moonshotlearning.org/';
+    // if we are in development, links are to localhost
+    if (!process.env.NODE_ENV) {
+        moonshotUrl = 'http://localhost:8081/';
+    }
+
+    Users.findOne(query, function (err, user) {
+        let recipient = [user.email];
+        let subject = 'Verify email';
+        let content =
+            '<div style="font-size:15px;text-align:center;font-family: Arial, sans-serif;color:#686868">'
+                + '<a href="' + moonshotUrl + '" style="color:#00c3ff"><img alt="Moonshot Logo" style="height:100px;margin-bottom:20px"src="https://image.ibb.co/iAchLn/Official_Logo_Blue.png"/></a><br/>'
+                    + '<div style="text-align:justify;width:80%;margin-left:10%;">'
+                    + '<span style="margin-bottom:20px;display:inline-block;">Thank you for joining Moonshot! To get going on your pathways, learning new skills, and building your profile for employers, please <a href="' + moonshotUrl + 'verifyEmail?token=' + user.emailVerificationToken + '">verify your account</a>.</span><br/>'
+                    + '<span style="display:inline-block;">If you have any questions or concerns or if you just want to talk about the weather, please feel free to email us at <a href="mailto:Support@MoonshotLearning.org">Support@MoonshotLearning.org</a>.</span><br/>'
+                    + '</div>'
+                + '<a style="display:inline-block;height:28px;width:170px;font-size:18px;border:2px solid #00d2ff;color:#00d2ff;padding:10px 5px 0px;text-decoration:none;margin:20px;" href="' + moonshotUrl + 'verifyEmail?token='
+                + user.emailVerificationToken
+                + '">VERIFY ACCOUNT</a>'
+                + '<div style="text-align:left;width:80%;margin-left:10%;">'
+                    + '<span style="margin-bottom:20px;display:inline-block;">On behalf of the Moonshot Team, we welcome you to our family and look forward to helping you pave your future and shoot for the stars.</span><br/>'
+                    + '<div style="font-size:10px; text-align:center; color:#C8C8C8; margin-bottom:30px;">'
+                    + '<i>Moonshot Learning, Inc.<br/><a href="" style="text-decoration:none;color:#D8D8D8;">1261 Meadow Sweet Dr<br/>Madison, WI 53719</a>.<br/>'
+                    + '<a style="color:#C8C8C8; margin-top:20px;" href="' + moonshotUrl + 'unsubscribe?email=' + user.email + '">Opt-out of future messages.</a></i>'
+                    + '</div>'
+                + '</div>'
+            + '</div>';
+
+        const sendFrom = "Moonshot";
+        sendEmail(recipient, subject, content, sendFrom, undefined, function (success, msg) {
+            if (success) {
+                res.json(msg);
+            } else {
+                res.status(500).send(msg);
+            }
+        })
     });
 }
