+++ src/actions/usersActions.js
@@ -139,7 +139,7 @@ export function postUser(user) {
             // user successfully posted
             .then(function(response) {
                 // send verification email
-                axios.post("/api/sendVerificationEmail", {email: user.email})
+                axios.post("/api/candidate/sendVerificationEmail", {email: user.email})
                     // successfully sent verification email
                     .then(function(emailResponse) {
                         dispatch({type:"POST_USER"});
@@ -430,22 +430,6 @@ export function comingSoon(user, signedIn){
             })
             .catch(function(err) {
                 dispatch({type:"FOR_BUSINESS", notification: {message: "Error sending email", type: "errorHeader"}})
-            })
-    }
-}
-
-// Send an email when a student registers for a pathway
-export function registerForPathway(user) {
-    return function(dispatch) {
-        dispatch({type: "REGISTER_FOR_PATHWAY_REQUESTED"});
-
-        axios.post("/api/user/registerForPathway", user)
-            .then(function(response) {
-                window.scrollTo(0, 0);
-                dispatch({type:"REGISTER_FOR_PATHWAY", notification: {message:response.data, type:"infoHeader"}});
-            })
-            .catch(function(err) {
-                dispatch({type:"REGISTER_FOR_PATHWAY", notification: {message: "Error sending email and registering, please try again.", type: "errorHeader"}})
             })
     }
 }
