+++ src/reducers/usersReducers.js
@@ -126,16 +126,6 @@ export function usersReducers(state = initialState, action) {
                 ...state, notification: action.notification, loadingSomething: false
             };
             break;
-        case "REGISTER_FOR_PATHWAY":
-            return {
-                ...state, notification: action.notification, loadingEmail: false
-            };
-            break;
-        case "REGISTER_FOR_PATHWAY_REQUESTED":
-            return {
-                ...state, loadingEmail: true
-            };
-            break;
         case "UPDATE_CURRENT_SUBSTEP":
             const subStep = {...action.payload, pathwayId: action.pathwayId}
             return {
